Improved Jumping Behaviour/Moving Platform
Jumping Behaviour used to be quite sluggish as often playtesters would press the jump check before landing on the platform, now player will jump on holding down of the jump key rather than key press down; provided they are not falling anymore.
Moving Platform used to add rigidbodys that are not ontop of the platform (whether on contact from the bottom or the side), added additional checks on hit normals to improve behaviour.